Rename dc_mimefactory.rs -> mimefactory.rs
Adjust imports as needed.
