+++ src/job.rs
@@ -10,13 +10,13 @@ use crate::configure::*;
 use crate::constants::*;
 use crate::context::Context;
 use crate::dc_imex::*;
-use crate::dc_mimefactory::*;
 use crate::dc_tools::*;
 use crate::events::Event;
 use crate::imap::*;
 use crate::location;
 use crate::login_param::LoginParam;
 use crate::message::*;
+use crate::mimefactory::*;
 use crate::param::*;
 use crate::sql;
 use crate::x::*;
